requirements.txt
anyio==4.7.0
asttokens==3.0.0
attrs==24.2.0
blessed==1.20.0
certifi==2024.8.30
charset-normalizer==3.4.0
codefind==0.1.7
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
fake-useragent==2.0.3
giving==0.4.3
h11==0.14.0
helium==5.1.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jurigged==0.6.0
markdown-it-py==3.0.0
mdurl==0.1.2
outcome==1.3.0.post0
ovld==0.4.6
packaging==24.2
psutil==6.1.0
Pygments==2.18.0
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==21.9
reactivex==4.0.4
requests==2.32.3
rich==13.9.4
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
urllib3==2.2.3
varname==0.13.5
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
